tests: replaced node_operations fuzzy test with scaled random node ops test
